Added geocoder to return search results
It gives results that aren't close to current location. It brings up a
dialog box with the results. I am trying to add lat/lng boundaries to
create a five mile radius but it quits the app with a
NullPointerException everytime I try to search with these boundaries.
Worked fine before I added boundaries.